Astarte.Flow.Blocks.MapSplitter (astarte_flow v0.1.0)

Breaks a map message into several messages, one for each map item.

Functions

split_map(message, config)

Specs

split_map(Astarte.Flow.Message.t(), Astarte.Flow.Blocks.MapSplitter.Config.t()) ::
  {:ok, [Astarte.Flow.Message.t()]} | {:error, reason :: term()}

Split a map message into a different message for each key.

start_link(opts)

Specs

start_link([Astarte.Flow.Blocks.MapSplitter.Config.option()]) ::
  GenServer.on_start()

Starts the MapSplitter.

Options

:key_action - the action to apply to the message key, such as :none, {:replace, delimiter}, {:append, delimiter} and {:prepend, delimiter}.

:fallback_action - fallback action that is performed when a message does not have map type, such as :discard, {:replace_key, new_key} and :pass_through.