Astarte.Flow.Blocks.Sorter.Config (astarte_flow v0.1.0)

Types

option()

@type option() :: {:deduplicate, boolean()} | {:delay_ms, pos_integer()}

t()

@type t() :: %Astarte.Flow.Blocks.Sorter.Config{delay_us: term(), policy: term()}

Functions

from_keyword(kl)

@spec from_keyword([option()]) :: {:ok, t()}

Initialize config from a keyword list.

Options

:deduplicate - true when messages deduplication is enabled, otherwise false. Two messages are duplicate when they have same timestamp and value (any other field is ignored).

:delay_ms - the amount of time the message is kept for reorder and deduplicate operations.