Astarte.Flow.Blocks.JsonPathMapper.JsonTemplate (astarte_flow v0.1.0)

Implements a simple logic less templating system that allows to write map templates with a syntax that resembles mustache syntax, that can be used for JSON templating.

Types

compiled_item()
(opaque)

@opaque compiled_item()

compiled_template()
(opaque)

@opaque compiled_template()

rendered_template()

@type rendered_template() ::
  template_base_type()
  | [rendered_template()]
  | %{optional(String.t()) => rendered_template()}

template()

@type template() ::
  template_base_type() | [template()] | %{optional(String.t()) => template()}

template_base_type()

@type template_base_type() :: String.t() | number() | boolean() | nil

Functions

compile_template(template_list)

@spec compile_template(template()) :: {:ok, compiled_template()} | {:error, any()}

render(template, input)

@spec render(compiled_template(), any()) ::
  {:ok, rendered_template()} | {:error, any()}