Astarte.FlowWeb.Router.Helpers (astarte_flow v0.1.0)

Module with named helpers generated from Astarte.FlowWeb.Router.

Functions

block_path(conn_or_endpoint, action, realm)

block_path(conn_or_endpoint, action, realm, params)

block_path(conn_or_endpoint, action, realm, name, params)

block_url(conn_or_endpoint, action, realm)

block_url(conn_or_endpoint, action, realm, params)

block_url(conn_or_endpoint, action, realm, name, params)

flow_path(conn_or_endpoint, action, realm)

flow_path(conn_or_endpoint, action, realm, params)

flow_path(conn_or_endpoint, action, realm, name, params)

flow_url(conn_or_endpoint, action, realm)

flow_url(conn_or_endpoint, action, realm, params)

flow_url(conn_or_endpoint, action, realm, name, params)

path(data, path)

Generates the path information including any necessary prefix.

pipeline_path(conn_or_endpoint, action, realm)

pipeline_path(conn_or_endpoint, action, realm, params)

pipeline_path(conn_or_endpoint, action, realm, name, params)

pipeline_url(conn_or_endpoint, action, realm)

pipeline_url(conn_or_endpoint, action, realm, params)

pipeline_url(conn_or_endpoint, action, realm, name, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.